hoan thien cac class staff

+++ src/main/java/com/se1858/group4/Land_Auction_SWP391/controller/NewsWriterController.java
@@ -1,16 +1,15 @@
 package com.se1858.group4.Land_Auction_SWP391.controller;
 
-import com.se1858.group4.Land_Auction_SWP391.dto.CustomerDTO;
 import com.se1858.group4.Land_Auction_SWP391.dto.NewsDTO;
 import com.se1858.group4.Land_Auction_SWP391.dto.StaffDTO;
 import com.se1858.group4.Land_Auction_SWP391.entity.Account;
-import com.se1858.group4.Land_Auction_SWP391.entity.Asset;
 import com.se1858.group4.Land_Auction_SWP391.entity.News;
 import com.se1858.group4.Land_Auction_SWP391.entity.TagForNews;
 import com.se1858.group4.Land_Auction_SWP391.security.UserDetailsService;
 import com.se1858.group4.Land_Auction_SWP391.service.NewsService;
 import com.se1858.group4.Land_Auction_SWP391.service.TagForNewsService;
 import com.se1858.group4.Land_Auction_SWP391.utility.FileUploadUtil;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -26,21 +25,19 @@ public class NewsWriterController {
     private NewsService newsService;
     private TagForNewsService tagForNewsService;
     private FileUploadUtil uploadFile;
-    @Autowired
     private UserDetailsService userDetailsService;
 
-
     @Autowired
-    public NewsWriterController(NewsService newsService, TagForNewsService tagForNewsService, FileUploadUtil uploadFile) {
+    public NewsWriterController(NewsService newsService, TagForNewsService tagForNewsService, FileUploadUtil uploadFile, UserDetailsService userDetailsService) {
         this.newsService = newsService;
         this.tagForNewsService = tagForNewsService;
         this.uploadFile = uploadFile;
+        this.userDetailsService = userDetailsService;
     }
     @GetMapping("/dashboard")
     public String dashboard( Model model) {
         return "newsWriter/Dashboard";
     }
-
 
     @GetMapping("/profile")
     public String showProfile(Model model) {
@@ -54,8 +51,6 @@ public class NewsWriterController {
         return "/newsWriter/profile";
     }
 
-
-
     @GetMapping("/create_news")
     public String createNews(Model model) {
         model.addAttribute("newsDTO", new NewsDTO());
@@ -65,15 +60,17 @@ public class NewsWriterController {
 
     @PostMapping("/add")
     public String addNews(@ModelAttribute("newsDTO") NewsDTO newsDTO,
-                          @RequestParam("selectedTags") List<Integer> selectedTags,
+                          @RequestParam(value = "selectedTags", required = false) List<Integer> selectedTags,
                           @RequestParam("images") List<MultipartFile> images,
                           RedirectAttributes redirectAttributes) {
         News news = newsDTO.getNews();
         uploadFile.UploadImagesForNews(images, news);
-        for (Integer tagId : selectedTags) {
-            TagForNews tag = tagForNewsService.getTagForNewsById(tagId);
-            news.addTag(tag);
-            tag.addNews(news);
+        if(selectedTags != null) {
+            for (Integer tagId : selectedTags) {
+                TagForNews tag = tagForNewsService.getTagForNewsById(tagId);
+                news.addTag(tag);
+                tag.addNews(news);
+            }
         }
         newsService.save(news);
         redirectAttributes.addAttribute("message", "The news was created successfully");
@@ -91,7 +88,8 @@ public class NewsWriterController {
 
     @GetMapping("/get_own_news_list")
     public String getOwnNews(Model model) {
-        List<News> list = newsService.getAllNews();
+        Account this_user = userDetailsService.accountAuthenticated();
+        List<News> list = newsService.getAllNewsByAuthorId(this_user.getAccountId());
         model.addAttribute("listNews", list);
         model.addAttribute("pageTitle", "View my own news");
         model.addAttribute("deletePermission", "true");
@@ -99,8 +97,16 @@ public class NewsWriterController {
     }
 
     @GetMapping("/viewDetail")
-    public String getNewsById(@RequestParam("newsId") int newsId, Model model) {
+    public String getNewsById(@RequestParam("newsId") int newsId, Model model, HttpServletRequest request) {
+        String previousUrl = request.getHeader("Referer");
+        System.out.println(previousUrl);
+        if (newsId <= 0) {
+            return "redirect:/news_writer/get_all_news_list";
+        }
         News news = newsService.getNewsById(newsId);
+        if(news==null){
+            return "redirect:/news_writer/get_all_news_list";
+        }
         model.addAttribute("news", news);
         return "newsWriter/NewsDetail";
     }
@@ -108,7 +114,13 @@ public class NewsWriterController {
     @GetMapping("/deleteNews")
     public String deleteNews(@RequestParam("newsId") int newsId) {
         Account account = userDetailsService.accountAuthenticated();
+        if (newsId <= 0) {
+            return "redirect:/news_writer/get_own_news_list";
+        }
         News news = newsService.getNewsById(newsId);
+        if(news==null){
+            return "redirect:/news_writer/get_own_news_list";
+        }
         if(news.getStaff().getAccountId()==account.getAccountId()) {
             uploadFile.deleteFile(news.getCover_photo().getPath());
             newsService.deleteNewsById(newsId);
